DRY transformGuildCreate
